requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic==2.11.7
python-dotenv==1.1.1
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
xgboost==3.0.4
joblib==1.5.2
ortools==9.14.6206
redis==6.4.0
celery==5.5.3
pytest==8.4.1
httpx==0.28.1
ruff==0.12.10
